+++ src/core.js
@@ -155,6 +155,12 @@ define(function (require, exports, module) {
                 skipHidden: self.get('skipHidden')
             }, cfg);
 
+            // 当 item 初始化的 element 为 selector 字符串时
+            // 默认到 validator.element 下去找
+            if (typeof cfg.element === 'string') {
+                cfg.element = this.element.find(cfg.element);
+            }
+
             if (!$(cfg.element).length) {
                 if (cfg.failSilently) {
                     return self;